Split comparison table rows into new component
