+++ webview/src/plots/components/comparisonTable/ComparisonTable.tsx
@@ -1,27 +1,18 @@
 import { ComparisonPlots, Revision } from 'dvc/src/plots/webview/contract'
 import { reorderObjectList } from 'dvc/src/util/array'
-import React, {
-  useState,
-  useRef,
-  useEffect,
-  useCallback,
-  createRef
-} from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import React, { useState, useRef, useEffect, useCallback } from 'react'
+import { useSelector } from 'react-redux'
 import { MessageFromWebviewType } from 'dvc/src/webview/contract'
-import { ComparisonTableRow } from './ComparisonTableRow'
 import {
   ComparisonTableColumn,
   ComparisonTableHead
 } from './ComparisonTableHead'
-import { RowDropTarget } from './RowDropTarget'
-import { changeRowHeight, DEFAULT_ROW_HEIGHT } from './comparisonTableSlice'
+import { ComparisionTableRows } from './ComparisonTableRows'
 import plotsStyles from '../styles.module.scss'
 import { withScale } from '../../../util/styles'
 import { sendMessage } from '../../../shared/vscode'
 import { PlotsState } from '../../store'
 import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
-import { DragDropContainer } from '../../../shared/components/dragDrop/DragDropContainer'
 
 export const ComparisonTable: React.FC = () => {
   const { plots } = useSelector((state: PlotsState) => state.comparison)
@@ -33,7 +24,6 @@ export const ComparisonTable: React.FC = () => {
   const pinnedColumn = useRef('')
   const [columns, setColumns] = useState<ComparisonTableColumn[]>([])
   const [comparisonPlots, setComparisonPlots] = useState<ComparisonPlots>([])
-  const [rowsOrder, setRowsOrder] = useState<string[]>([])
 
   const isPinned = (column: ComparisonTableColumn): boolean =>
     column.revision === pinnedColumn.current
@@ -42,9 +32,6 @@ export const ComparisonTable: React.FC = () => {
     () => revisions?.find(isPinned) || null,
     [revisions]
   )
-
-  const dispatch = useDispatch()
-  const firstRowRef = createRef<HTMLTableSectionElement>()
 
   useEffect(
     // eslint-disable-next-line sonarjs/cognitive-complexity
@@ -68,14 +55,7 @@ export const ComparisonTable: React.FC = () => {
 
   useEffect(() => {
     setComparisonPlots(plots)
-    setRowsOrder(plots.map(({ path }) => path))
   }, [plots])
-
-  const onLayoutChange = () => {
-    const firstRowHeight =
-      firstRowRef.current?.getBoundingClientRect().height || DEFAULT_ROW_HEIGHT
-    dispatch(changeRowHeight(firstRowHeight))
-  }
 
   if (!plots || plots.length === 0) {
     return <EmptyState isFullScreen={false}>No Images to Compare</EmptyState>
@@ -103,27 +83,6 @@ export const ComparisonTable: React.FC = () => {
     )
   }
 
-  const rows = rowsOrder.map((path, i) => {
-    const plot = comparisonPlots.find(p => p.path === path)
-    if (!plot) {
-      return
-    }
-    const revs = plot.revisions
-    return (
-      <tbody key={path} id={path} ref={i === 0 ? firstRowRef : undefined}>
-        <ComparisonTableRow
-          path={path}
-          plots={columns.map(column => ({
-            ...revs[column.revision],
-            revision: column.revision
-          }))}
-          nbColumns={columns.length}
-          pinnedColumn={pinnedColumn.current}
-        />
-      </tbody>
-    )
-  })
-
   return (
     <table
       className={plotsStyles.comparisonTable}
@@ -135,14 +94,10 @@ export const ComparisonTable: React.FC = () => {
         setColumnsOrder={setColumnsOrder}
         setPinnedColumn={changePinnedColumn}
       />
-      <DragDropContainer
-        items={rows as JSX.Element[]}
-        order={rowsOrder}
-        setOrder={setRowsOrder}
-        group="comparison-table"
-        dropTarget={<RowDropTarget colSpan={columns.length} />}
-        onLayoutChange={onLayoutChange}
-        vertical
+      <ComparisionTableRows
+        plots={comparisonPlots}
+        columns={columns}
+        pinnedColumn={pinnedColumn.current}
       />
     </table>
   )
